requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
alembic==1.9.2
psycopg2==2.9.5
python-dotenv==0.21.1
FastAPI-SQLAlchemy==0.2.1